+++ clarte360-emargements-v1.0.0/tests/test_v2_candidate.py
@@ -22,4 +22,4 @@ def test_candidate_transferable_pdf_branding():
 
 def test_candidate_version_is_rc():
     from branding import APP_VERSION
-    assert APP_VERSION=='3.0.0-I9-J2C-PIP-LIAISON-I-CRM0-RC1'
+    assert APP_VERSION=='3.0.0-I9-J2C-PIP-LIAISON-I-CRM0-RC2'